Stage #2

+++ Shared Bills Splitter/task/src/splitter/model/Transaction.java
@@ -1,14 +1,15 @@
 package splitter.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Transaction {
     private final LocalDate date;
     private final Person creditor;
     private final Person debtor;
-    private final int sum;
+    private final BigDecimal sum;
 
-    public Transaction(LocalDate date, Person from, Person to, int sum) {
+    public Transaction(LocalDate date, Person from, Person to, BigDecimal sum) {
         this.date = date;
         this.creditor = from;
         this.debtor = to;
@@ -27,7 +28,7 @@ public class Transaction {
         return debtor;
     }
 
-    public int getSum() {
+    public BigDecimal getSum() {
         return sum;
     }
 }
